Moving data transmission and delay calculation into module .py
